refactor: divide components and add public api for modules
