+++ src/app/styled/index.ts
@@ -1,12 +1,5 @@
-import styled, { createGlobalStyle } from 'styled-components';
+import { createGlobalStyle } from 'styled-components';
 import { LIGHT_THEME } from '@admiral-ds/react-ui';
-
-export const InputsStyled = styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-    width: 400px;
-`;
 
 export const BodyGlobal = createGlobalStyle`
     body {
