+++ src/layouts/Layout.tsx
@@ -1,10 +1,9 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from '../components/Header';
 import { LayoutStyled } from './Layout.styled';
-import { BodyGlobal } from '../styled';
+import { Header } from '../modules/user';
+import { BodyGlobal } from '../app/styled';
 
 export function Layout() {
-
     return (
         <LayoutStyled>
             <Header />
