+++ src/main.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
 import { ThemeProvider } from 'styled-components';
 import { DropdownProvider } from '@admiral-ds/react-ui';
-import { CUSTOM_THEME } from './modules/shared/themes/CUSTOM_THEME';
+import { CUSTOM_THEME } from './app/themes/CUSTOM_THEME';
 import { Provider } from 'react-redux';
-import ErrorBoundary from './modules/shared/components/ErrorBoundary';
+import ErrorBoundary from './app/ErrorBoundary';
 import { RouterProvider } from 'react-router-dom';
 import { routesConfig } from './app/routes/routesConfig';
 import { GlobalFont } from './app/GlobalFont';
