--- src/modules/shared/components/Header.styled.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-import { LIGHT_THEME } from '@admiral-ds/react-ui';
-
-export const HeaderStyled = styled.header`
-    display: flex;
-    height: 55px;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 20px;
-    background-color: ${LIGHT_THEME.color['Neutral/Neutral 00']};
-`;
