+++ src/modules/user/components/AvatarUser/AvatarUser.styled.ts
