+++ src/modules/user/components/AvatarUser/index.ts
@@ -1,0 +1,1 @@
+export { AvatarUser } from './AvatarUser';
