+++ src/modules/user/components/RegistrationForm/RegistrationForm.styled.ts
