+++ src/modules/user/components/index.ts
@@ -1,0 +1,1 @@
+export { Header } from './Header';
