+++ src/modules/user/hooks/useAuthLogOut.ts
@@ -1,30 +1,38 @@
+import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { useLogoutMutation } from '../services/apiService';
 import { RoutesEnum } from '../../../app/routes/RoutesEnum';
-import { useDispatch } from 'react-redux';
 import { logOut as logoutAction } from '../slice/authSlice';
-import { setShowLogout } from '../../shared/slice/uiSlice';
+import { setShowLogout, selectLogOut } from '../slice/uiSlice';
+import { selectUser } from '../slice/authSlice';
 
-export const useLogOut = () => {
-    const navigate = useNavigate();
-    const [logout, { isLoading }] = useLogoutMutation();
+export const useAuthLogOut = () => {
     const dispatch = useDispatch();
+    const navigate = useNavigate();
+
+    const user = useSelector(selectUser);
+    const showLogout = useSelector(selectLogOut);
+
+    const [logout, { isLoading }] = useLogoutMutation();
 
     const handleLogout = async () => {
         try {
-            await logout({})
+            await logout({});
             localStorage.removeItem('user');
             dispatch(logoutAction());
             navigate(RoutesEnum.ROOT);
         } catch (error) {
             console.error(error);
         } finally {
-            dispatch(setShowLogout(false))
+            dispatch(setShowLogout(false));
         }
     };
 
     return {
         handleLogout,
-        isLoading
+        isLoading,
+        dispatch,
+        user,
+        showLogout,
     };
 };
