+++ src/modules/user/hooks/useLogOut.tsx
@@ -1,39 +1,30 @@
-import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { useLogoutMutation } from '../services/apiService';
 import { RoutesEnum } from '../../../app/routes/RoutesEnum';
+import { useDispatch } from 'react-redux';
 import { logOut as logoutAction } from '../slice/authSlice';
-import { setShowLogout, selectLogOut } from '../../shared/slice/uiSlice';
-import { selectUser } from '../slice/authSlice'
+import { setShowLogout } from '../slice/uiSlice';
 
-export const useAuthLogOut = () => {
-    const dispatch = useDispatch();
+export const useLogOut = () => {
     const navigate = useNavigate();
-
-    const user = useSelector(selectUser);
-    const showLogout = useSelector(selectLogOut);
-
     const [logout, { isLoading }] = useLogoutMutation();
+    const dispatch = useDispatch();
 
     const handleLogout = async () => {
         try {
-            await logout({})
+            await logout({});
             localStorage.removeItem('user');
             dispatch(logoutAction());
             navigate(RoutesEnum.ROOT);
         } catch (error) {
             console.error(error);
         } finally {
-            dispatch(setShowLogout(false))
+            dispatch(setShowLogout(false));
         }
     };
 
     return {
         handleLogout,
         isLoading,
-        dispatch,
-        user,
-        showLogout,
-    }
-}
-
+    };
+};
