+++ src/shared/components/ActionBar/index.ts
@@ -1,0 +1,1 @@
+export { ActionBar } from './ActionBar';
