+++ src/shared/components/index.ts
@@ -1,0 +1,2 @@
+export { ActionBar } from './ActionBar';
+export { AvatarUser } from '../../modules/user/components/AvatarUser';
