+++ src/shared/index.ts
@@ -1,0 +1,2 @@
+export { ActionBar, AvatarUser } from './components';
+export { InputsStyled } from './styled';
