+++ src/shared/styled/index.ts
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const InputsStyled = styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    width: 400px;
+`;
